+++ .github/CHANGELOG.md
@@ -29,6 +29,8 @@
 1. [A380X/MFD] Fixed DIRECT TO selection of active waypoint does nothing - @sognodelx (Sven Gross)
 1. [A380X/FLIGHT MODEL] Fix for stall and auto-rotation issues- @donstim (donbikes)
 1. [A380X/MODEL] Optimized 4K textures - @Repsol
+1. [A380X/EFB] Added Minimum Runway Length Option - @ksleungac (ksleungac)
+1. [A380X/EFIS] Read Minimum Runway Length setting from EFB to filter airports for EFIS/ND - @ksleungac (ksleungac)
 
 ## 0.12.0